Move MoveItErrorCode class to a standalone file
